narcisss: install_linguas don't install empty list
* in some combination of IMAGE_LINGUAS installed packages and available translations it's normal to get empty intersection in pending-locale-packages
* don't call opkg install in this case, because it will fail

From http://cgit.openembedded.org/cgit.cgi/openembedded/commit/?id=e38c2c1286f25518770d6c668dc43939476f312d by Martin Jansa

+++ scripts/install-package.sh
@@ -64,10 +64,12 @@ bin/opkg-cl ${CACHE} -o ${TARGET_DIR} -f ${TARGET_DIR}/etc/opkg.conf list | awk 
 
 cat ${TARGET_DIR}/tmp/wanted-locale-packages ${TARGET_DIR}/tmp/available-locale-packages | sort | uniq -d > ${TARGET_DIR}/tmp/pending-locale-packages
 
-for i in $(cat ${TARGET_DIR}/tmp/pending-locale-packages) ; do
-	echo "running: bin/opkg-cl ${CACHE} -o ${TARGET_DIR} -f ${TARGET_DIR}/etc/opkg.conf -nodeps install $i"
-	bin/opkg-cl ${CACHE} -o ${TARGET_DIR} -f ${TARGET_DIR}/etc/opkg.conf -nodeps install $i
-done
+if [ -s ${TARGET_DIR}/tmp/pending-locale-packages ] ; then
+	for i in $(cat ${TARGET_DIR}/tmp/pending-locale-packages) ; do
+		echo "running: bin/opkg-cl ${CACHE} -o ${TARGET_DIR} -f ${TARGET_DIR}/etc/opkg.conf -nodeps install $i"
+		bin/opkg-cl ${CACHE} -o ${TARGET_DIR} -f ${TARGET_DIR}/etc/opkg.conf -nodeps install $i
+	done
+fi
 
 echo "<div id=\"imgsize\">" $(du ${TARGET_DIR} -hs) "</div>"
